+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Allow structured responses.
 - Allow yada or JSON schema.
 - Using `template` option only now.
+- Allow specifying model in client options.
+- Added debug feature.
 
 ## 0.3.0